Resolve several Windows compile warnings. (dm)
git-svn-id: svn://mielke.cc/main/brltty@7401 91a5dbb7-01b9-0310-9b5f-b28072856b6e

+++ Programs/brlapi_server.c
@@ -1818,7 +1818,7 @@ out:
 
 static void *establishSocket(void *arg)
 {
-  long int num = (intptr_t) arg;
+  intptr_t num = (intptr_t) arg;
   struct socketInfo *cinfo = &socketInfo[num];
 #ifndef __MINGW32__
   int res;
